fix: observer was not being removed
