+++ Sources/PagecallSDK/MiController.swift
@@ -145,8 +145,6 @@ class MiController: MediaController, SendTransportDelegate, ReceiveTransportDele
 
         sendTransport.delegate = self
         recvTransport.delegate = self
-
-        startHandlingInterruption()
     }
 
     func consume(data: Data, callback: (Error?) -> Void) {
@@ -218,7 +216,6 @@ class MiController: MediaController, SendTransportDelegate, ReceiveTransportDele
     }
 
     func dispose() {
-        stopHandlingInterruption()
         sendTransport.close()
         recvTransport.close()
     }
